backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pymysql==1.1.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
langchain==0.2.16
langchain-google-genai==1.0.10
